requirements.txt
asgiref==3.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-location-field==2.1.0
djangorestframework==3.12.4
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
ratelim==0.1.6
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
